fix issue with exported cli completion types
