+++ .changeset/hot-taxis-fry.md
@@ -1,0 +1,5 @@
+---
+"effect": patch
+---
+
+Fix issue with exported CLI `Completions` types
